CSS and API Additions Icons added logout button login form section view subjects faculty details section details Content Added svg editing
