+++ package.json
@@ -17,6 +17,7 @@
     "@mui/material": "^5.15.7",
     "@mui/x-data-grid": "^6.19.3",
     "@syncfusion/ej2-react-spreadsheet": "^24.2.4",
+    "multiselect-react-dropdown": "^2.0.25",
     "react": "^18.2.0",
     "react-datepicker": "^6.1.0",
     "react-dom": "^18.2.0",
