+++ src/components/Login/Content.jsx
@@ -1,0 +1,30 @@
+import React from "react";
+
+const Content = () => {
+  return (
+    <div className="top-[175px] left-[81px] h-2/4 w-[89vw] overflow-hidden text-justify font-roboto nl:h-[60vh] nl:w-screen lg:fixed sl:w-screen">
+      <div className="relative pt-8 w-[41vw] h-[43vh] nl:w-screen sl:w-screen">
+        <dv className="text-start nl:flex justify-center nl:items-start nl:text-center  sl:flex sl:flex-col sl:items-center sl:text-center ">
+          <div className="text-45xl text-gray sl:text-13xl mr-4">
+            Automation
+          </div>
+          <div className="-mt-7 text-45xl text-mediumslateblue-200 w-[516px]  nl:text-13x1 nl:w-1/2 nl:mt-0 sl:text-13xl">
+            for your Exams
+          </div>
+        </dv>
+        <div className="text-mini text-slategray-100 nl:mx-8  flex justify-center items-center sl:px-6 sl:text-Iphone">
+          Introducing the pinnacle of excellence in academic assessment: our
+          exclusive college exam automation platform. Immerse yourself in a
+          world of sophistication and precision as we elevate your result
+          management to an art form. Seamlessly capture and curate student marks
+          while our refined system orchestrates an impeccable evaluation,
+          leaving no room for compromise. Experience the epitome of efficiency
+          and opulence, where every achievement is meticulously crafted and
+          effortlessly showcased.
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Content;
