+++ src/components/footer/Footer.jsx
@@ -3,7 +3,7 @@ const Footer = () => {
     return (<div className="footer">
       <span>EduProctor</span>
       <div className="logotxt">
-        <img className= "logo" src="../../../public/aiktc.jpg"/>
+        <img className= "logo" src="../../../aiktc.jpg"/>
       <span>AIKTC Exam Automation Admin Dashboard</span>
       </div>
